Rename sat directories and operation
