+++ lib/ops/adaptive_threshold/index.js
@@ -8,7 +8,7 @@
 
 import RegisterOperation from '../../program/operation_register';
 import * as utils from '../../utils';
-import { sat } from '../summed_area_table';
+import { sat } from '../sat';
 import kernel from './kernel.glsl';
 
 /**
