+++ lib/ops/index.js
@@ -33,6 +33,6 @@ export { default as erode } from './erode';
 export { default as dilate } from './dilate';
 export { default as morphologyEx } from './morphology_ex';
 export { default as upsample } from './upsample';
-export { sat, sqsat } from './summed_area_table';
+export { sat, sqsat } from './sat';
 export { default as adaptiveThreshold } from './adaptive_threshold';
 export { sub, div, mult, add, subScalar, divScalar, multScalar, addScalar } from './math';
