+++ lib/ops/sat/index.js
@@ -58,7 +58,7 @@ const sumedAreaTableBase = (tSrc, passesPerAxis = 2, squared = false) => {
 };
 
 /**
- * @name SummedAreaTable
+ * @name Sat
  * @description
  *  A summed-area table operation is quickly and efficiently generate
  *  the sum of values in a rectangular subset of a grid.
@@ -74,7 +74,7 @@ export const sat = (tSrc, passesPerAxis = 2) =>
   sumedAreaTableBase(tSrc, passesPerAxis, false);
 
 /**
- * @name SquaredSummedAreaTable
+ * @name SqSat
  * @description
  *  A squared summed-area table operation is quickly and efficiently generate
  *  the sum of squared values in a rectangular subset of a grid.
