+++ lib/ops/sat/kernel_x_sq.glsl
